Fixing canvas undefined and module undefined errors + reverting Susan's changes
