+++ client/index.html
@@ -6,16 +6,13 @@
 <html lang="en">
     <head>
         <meta charset="UTF-8">
-        <!--set width of screen to device width, set initial zoom to 90%-->
         <meta name="viewport" content="width=device-width, initial-scale=0.9"> 
         <title>Turbo</title>
         <link rel='stylesheet' href='css/style.css'/>    
         
     </head>
     
-    <body id ="home">
-        
-        
+    <body id ="home">        
 
         <div class ="header">
             </div>
