+++ client/multiplayer.html
@@ -6,9 +6,11 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Multiplayer</title>
     <link rel='stylesheet' href='css/multiplayer.css'/>
-   <!-- <script src="./js/fuel.js"></script>    ***Temporarrily commenting this out***
-    <script src="./js/liveplayers.js"></script>-->
     <script src ="https://cdnjs.cloudflare.com/ajax/libs/socket.io/2.3.0/socket.io.js"></script>
+    <script src="js/liveplayers.js"></script>
+    <script src="js/fuel.js "></script>
+    
+
 
 </head>
 <body>
